chore:added jwt

+++ src/App.js
@@ -6,13 +6,14 @@ function App() {
   const [image, setImage] = useState("");
   const [result, setResult] = useState("");
   const [deleteMessage, setDeleteSuccess] = useState("");
+  const [token, setToken] = useState("");
 
   useEffect(() => {
     //   Calling get api
-    getProductDetails()
-    
-    //     making post api call. 
-    PostingContent()
+    getProductDetails();
+
+    //     making post api call.
+    PostingContent();
 
     //    To get image from api
     const getImage = () => {
@@ -26,12 +27,13 @@ function App() {
         setImage(base64);
       });
     };
-
     getImage();
+
+    //   User Token
+    CreateUserToken();
   }, []);
 
-
-  const getProductDetails=()=>{
+  const getProductDetails = () => {
     api
       .getallData()
       .then((response) => {
@@ -40,24 +42,32 @@ function App() {
       .catch((error) => {
         console.error(error);
       });
-  }
+  };
 
-  const PostingContent=async()=>{
+  const PostingContent = async () => {
     //  making post api call. here in body we are sending name:'GeekyAnt' so in backend we can access it through req.body.name
-    let response = await api.postingBody({name:'GeekyAnt'})
-    let responseTwo = await api.postingBodyTwo({name:'Sir'})
+    let response = await api.postingBody({ name: "GeekyAnt" });
+    let responseTwo = await api.postingBodyTwo({ name: "Sir" });
 
-    setResult(response.data +  responseTwo.data);
-  }
+    setResult(response.data + responseTwo.data);
+  };
 
   const deleteItem = async (params) => {
     try {
-      const deleteResult = await api.deleteItem(params);
-      setDeleteSuccess(deleteResult.data.status);
-      getProductDetails()
+      const userValidation = await api.getUserToken(token);
+      if (userValidation.status === 200) {
+        try {
+          const deleteResult = await api.deleteItem(params);
+          setDeleteSuccess(deleteResult.data.status);
+          getProductDetails();
+        } catch (error) {
+          console.error(error);
+        } finally {
+          setDeleteSuccess("");
+        }
+      }
     } catch (error) {
-    } finally {
-      setDeleteSuccess("");
+      alert("Something went wrong");
     }
   };
 
@@ -68,29 +78,31 @@ function App() {
       name: `Product ${token}`,
       description: `Description of Product ${token}`,
       price: Date.now(),
-      image:"https://media.geeksforgeeks.org/wp-content/uploads/20230728155224/images.jfif",
+      image:
+        "https://media.geeksforgeeks.org/wp-content/uploads/20230728155224/images.jfif",
     };
     const createdItemsSuccess = await api.createItem(payload);
-    getProductDetails()
+    getProductDetails();
     console.log(createdItemsSuccess);
   };
 
+  const updateItem = async (params) => {
+    const updatedItems = await api.updateItemTwo(params);
+    getProductDetails();
+  };
 
+  const GetItem = async (params) => {
+    const getSpecificItem = await api.getSpecificItem(params);
+    //                    OR
+    const getSpecificItemData = await api.getSpecificItemByParams(params);
+    //                      OR
+    const getSpecificItemWithQuery = await api.getSpecificItemByQuery(params);
+  };
 
-
-const updateItem= async(params)=>{
-  const updatedItems = await api.updateItemTwo(params)
-  getProductDetails()
-}
-
-const GetItem=async(params)=>{
-  const getSpecificItem= await api.getSpecificItem(params)
-  //                    OR
-  const getSpecificItemData= await api.getSpecificItemByParams(params)
-//                      OR
-  const  getSpecificItemWithQuery = await api.getSpecificItemByQuery(params)
-
-}
+  const CreateUserToken = async () => {
+    const userToken = await api.createUserToken();
+    setToken(userToken.data);
+  };
 
   return (
     <div className="App">
@@ -106,11 +118,29 @@ const GetItem=async(params)=>{
                 <img className="img" src={data.image} alt="img" />
                 <h1>{data.name}</h1>
                 <p>{data.description}</p>
-                <p>Rs. {"  "}{data.price}</p>
+                <p>
+                  Rs. {"  "}
+                  {data.price}
+                </p>
                 <div>
-                  <button onClick={() => deleteItem(data)} style={{marginLeft:'5px',marginRight:'5px'}}>Delete Item</button>
-                  <button onClick={()=>updateItem(data)} style={{marginLeft:'5px',marginRight:'5px'}}>Update Item</button>
-                  <button onClick={()=>GetItem(data)} style={{marginLeft:'5px',marginRight:'5px'}}>Get Item</button>
+                  <button
+                    onClick={() => deleteItem(data)}
+                    style={{ marginLeft: "5px", marginRight: "5px" }}
+                  >
+                    Delete Item
+                  </button>
+                  <button
+                    onClick={() => updateItem(data)}
+                    style={{ marginLeft: "5px", marginRight: "5px" }}
+                  >
+                    Update Item
+                  </button>
+                  <button
+                    onClick={() => GetItem(data)}
+                    style={{ marginLeft: "5px", marginRight: "5px" }}
+                  >
+                    Get Item
+                  </button>
                 </div>
               </div>
             );
@@ -118,7 +148,12 @@ const GetItem=async(params)=>{
         </div>
       }
       <div>
-        <img src={`data:;base64,${image}`} alt="Ref_Image" height='300px'  loading="eager"/>
+        <img
+          src={`data:;base64,${image}`}
+          alt="Ref_Image"
+          height="300px"
+          loading="eager"
+        />
       </div>
       <div>{deleteMessage && deleteMessage}</div>
     </div>
